wip: use cobra to handle cli subcommands
