+++ go.mod
@@ -3,10 +3,10 @@ module github.com/sgaunet/gitlab-token-expiration
 go 1.23.2
 
 require (
+	github.com/fatih/color v1.18.0
 	github.com/pterm/pterm v0.12.79
 	github.com/savioxavier/termlink v1.4.1
-	golang.org/x/sync v0.9.0
-	golang.org/x/time v0.8.0
+	github.com/spf13/cobra v1.8.1
 )
 
 require (
@@ -14,13 +14,14 @@ require (
 	atomicgo.dev/keyboard v0.2.9 // indirect
 	atomicgo.dev/schedule v0.1.0 // indirect
 	github.com/containerd/console v1.0.3 // indirect
-	github.com/fatih/color v1.18.0 // indirect
 	github.com/gookit/color v1.5.4 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/lithammer/fuzzysearch v1.1.8 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mattn/go-runewidth v0.0.15 // indirect
 	github.com/rivo/uniseg v0.4.4 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/xo/terminfo v0.0.0-20220910002029-abceb7e1c41e // indirect
 	golang.org/x/sys v0.25.0 // indirect
 	golang.org/x/term v0.16.0 // indirect
