+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/sgaunet/gitlab-token-expiration/cmd"
+
+func main() {
+	cmd.Execute()
+}
